Added removeCredit() and rollBack() Changed double to BigDecimal in Product, ProductDto model

+++ product-service/src/main/java/com/creditApp/model/Product.java
@@ -9,6 +9,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import java.math.BigDecimal;
 
 @Entity
 @Getter
@@ -20,6 +21,6 @@ public class Product {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
     private String name;
-    private double value;
+    private BigDecimal value;
     private String creditNumber;
 }
